move: service 테스트 패키지 이동
